archive(old file)/requirements.txt
numpy
pyaudio
wave
rsa
Pillow
matplotlib
pycryptodome
scipy
seaborn
pydub